More proper navigation approach
Standardised properties among the models.

USAGE

add NavigatableTrait to the class you wanna use it with

the class must provide a makeLink method that returns an array of the properties to be stored in the corresponding NavigationItem model.

class->navItem to fetch/create a unique instance of NavigationItem

NavMenu::byName(<name>)->add(<instance>)
tries to call the <instance>->navItem method, using the trait's relation

in alternative an instance of NavigationItem can be passed
NavMenu::byName(<name>)->add(<instance of NavigationItem>)