reshaped navigations, solidified code.
-> MakeLink methods now MUST accept $overrides = [] element and array_replace them.

Navigation Items now have a better data integrity. Duplicates are forbidden in multiple ways to ensure the instance remains the same. Navmenu has a stricter way to validate items in it. Every step of the NavItem creation now accepts an array of overrides.

Moved templates to their current assigned location. We might wanna have some more conversation on that subject