Started testing actions permission on models Authorize on models is back to work (as a secondary, in-controller way to authenticate user against an action) Better API on HasPermissions Trait, allowing for both static and instance calls ((through Static prefix on method, read note)
NOTE: working with __callStatic revealed to be time consuming as it conflicts with laravel's, but it's something we should go deeper on. calling a method in both static and instance mode would be very useful in this scenario
NOTE: website current is returning Website::admin() as fallback. Helps isolating tests and interacting through CLI. Not sure if acceptable